requirements.txt
PyQt6>=6.4.0
Pillow>=9.0.0
google-genai>=0.8.0
openai>=1.0.0
markdown>=3.4.0
pygments>=2.10.0
yt-dlp>=2023.0.0
PyPDF2>=3.0.0